Added utils file
